python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
openpyxl==3.11.0
numpy==1.26.2
pydantic==2.5.0
python-multipart==0.0.6
cors==1.0.1
pymysql==1.1.0
python-dotenv==1.0.0
